pyproject.toml
[tool.poetry]
name = "visemes-audio-generator"
version = "0.1.0"
description = ""
authors = ["Alejandro Sánchez Yalí <asanchezyali@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
azure-cognitiveservices-speech = "^1.32.1"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
